# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HiveTrust API",
    "version": "1.0.0",
    "description": "KYA Identity Verification, Trust Scoring & Parametric Insurance for AI Agents in the A2A Economy.\n\nAll endpoints require an `X-API-Key` header unless noted. Costs are billed per call and reflected in your operator account.",
    "contact": {
      "name": "HiveTrust Support",
      "url": "https://hivetrust.hiveagentiq.com",
      "email": "support@hiveagentiq.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://hivetrust.hiveagentiq.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "security": [
    { "ApiKeyAuth": [] }
  ],
...
  "paths": {
...
    "/v1/agents": {
      "post": {
        "tags": ["Identity"],
        "summary": "Register an agent",
        "description": "Creates a new agent identity, assigns a DID (`did:hive:<uuid>`), and stores the Ed25519 public key. Cost: $0.01.",
        "operationId": "registerAgent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterAgentRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agent created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Agent" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
    "/v1/agents/{agentId}": {
      "get": {
        "tags": ["Identity"],
        "summary": "Get agent profile",
        "operationId": "getAgent",
        "parameters": [{ "$ref": "#/components/parameters/agentId" }],
        "responses": {
          "200": {
            "description": "Agent profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Agent" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Identity"],
        "summary": "Update agent metadata",
        "description": "Updates mutable fields and creates an immutable version snapshot. Cost: $0.01.",
        "operationId": "updateAgent",
        "parameters": [{ "$ref": "#/components/parameters/agentId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateAgentRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated agent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Agent" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Identity"],
        "summary": "Deactivate agent",
        "description": "Marks the agent as inactive and adds it to the revocation registry.",
        "operationId": "deactivateAgent",
        "parameters": [{ "$ref": "#/components/parameters/agentId" }],
        "responses": {
          "200": {
            "description": "Agent deactivated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessMessage" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/agents/{agentId}/credentials": {
      "post": {
        "tags": ["Credentials"],
        "summary": "Issue a credential",
        "description": "Issues a signed W3C Verifiable Credential for the agent. Cost: $0.05.",
        "operationId": "issueCredential",
        "parameters": [{ "$ref": "#/components/parameters/agentId" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/IssueCredentialRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Credential issued",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Credential" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "get": {
        "tags": ["Credentials"],
        "summary": "List agent credentials",
        "operationId": "listCredentials",
        "parameters": [
          { "$ref": "#/components/parameters/agentId" },
          { "$ref": "#/components/parameters/limit" },
          { "$ref": "#/components/parameters/offset" }
        ],
        "responses": {
          "200": {
            "description": "List of credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "credentials": { "type": "array", "items": { "$ref": "#/components/schemas/Credential" } },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/agents/{agentId}/credentials/{credentialId}": {
      "delete": {
        "tags": ["Credentials"],
        "summary": "Revoke a credential",
        "operationId": "revokeCredential",
        "parameters": [
          { "$ref": "#/components/parameters/agentId" },
          { "$ref": "#/components/parameters/credentialId" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": { "type": "string", "description": "Reason for revocation" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Credential revoked",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessMessage" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/v1/verify/credential": {
      "post": {
        "tags": ["Credentials"],
        "summary": "Verify a credential",
        "description": "Checks signature validity, expiry, and revocation status. Cost: $0.01.",
        "operationId": "verifyCredential",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/VerifyCredentialRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VerifyCredentialResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/agents/{agentId}/score": {
      "get": {
        "tags": ["Trust Score"],
        "summary": "Get trust score",
        "description": "Returns the current composite score and all five pillar sub-scores. Cost: $0.01.",
        "operationId": "getTrustScore",
        "parameters": [{ "$ref": "#/components/parameters/agentId" }],
        "responses": {
          "200": {
            "description": "Trust score",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TrustScore" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/v1/telemetry/ingest": {
      "post": {
        "tags": ["Telemetry"],
        "summary": "Ingest behavioural events",
        "description": "Bulk event ingestion (up to 1 000 events per call). Events feed the reputation engine. Cost: $0.001/event.",
        "operationId": "ingestTelemetry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TelemetryIngestRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Events accepted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TelemetryIngestResponse" }
              }
            }
          }
        }
      }
    },
...
    "/v1/insurance/quote": {
      "post": {
        "tags": ["Insurance"],
        "summary": "Get insurance quote",
        "description": "Calculates a dynamic premium based on both agents' trust scores and transaction value.",
        "operationId": "getInsuranceQuote",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InsuranceQuoteRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Insurance quote",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InsuranceQuoteResponse" }
              }
            }
          }
        }
      }
    },
    "/v1/insurance/bind": {
      "post": {
        "tags": ["Insurance"],
        "summary": "Bind insurance policy",
        "description": "Binds the quoted policy and deploys USDC escrow on Base L2. Take rate: 1.5% of insured value.",
        "operationId": "bindInsurance",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InsuranceBindRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Policy bound",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InsurancePolicy" }
              }
            }
          }
        }
      }
    },
...
    "/v1/insurance/claims": {
      "post": {
        "tags": ["Insurance"],
        "summary": "File a claim",
        "operationId": "fileClaim",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FileClaimRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Claim filed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Claim" }
              }
            }
          }
        }
      }
    },
...
    "/v1/disputes": {
      "post": {
        "tags": ["Disputes"],
        "summary": "File a dispute",
        "operationId": "fileDispute",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/FileDisputeRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dispute filed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Dispute" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key obtained from the HiveTrust dashboard or via the seed script for testing."
      }
    },
    "parameters": {
...
      "limit": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
...
    },
...
    "schemas": {
...
      "RegisterAgentRequest": {
        "type": "object",
        "required": ["name", "public_key", "endpoint_url"],
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "operator_name": { "type": "string" },
          "operator_contact": { "type": "string" },
          "model_fingerprint": { "type": "string" },
          "public_key": { "type": "string", "description": "Ed25519 public key, base58 encoded" },
          "endpoint_url": { "type": "string", "format": "uri" },
          "capability_manifest": { "type": "object" }
        }
      },
...
    }
  }
}